Move our headers into truffleruby, and try to reduce the number of headers called ruby.h
