+++ lib/cext/include/ruby/config.h
@@ -3,7 +3,7 @@
 // $ CC=clang ./configure
 // $ find . -name config.h
 #ifdef __APPLE__
-#include <truffle/config_darwin.h>
+#include <truffleruby/config_darwin.h>
 #else
-#include <truffle/config_linux.h>
+#include <truffleruby/config_linux.h>
 #endif
